Added new ability types
